+++ pkg/grafana/datasources.go
@@ -20,7 +20,7 @@ func (s *Service) ConfigureDatasource(ctx context.Context, organization Organiza
 	logger := log.FromContext(ctx)
 
 	// Generate the desired datasources for the organization
-	desiredDatasources := s.generateDatasources(ctx, organization)
+	desiredDatasources := s.generateDatasources(organization)
 
 	// Configure Grafana client to use the correct organization
 	currentOrgID := s.grafanaClient.OrgID()
@@ -45,7 +45,7 @@ func (s *Service) ConfigureDatasource(ctx context.Context, organization Organiza
 			desiredDatasource := desiredDatasources[index]
 
 			logger.Info("updating datasource", "datasource", desiredDatasource.UID)
-			desiredDatasource, err = s.updateDatasource(ctx, desiredDatasource)
+			desiredDatasource, err = s.updateDatasource(desiredDatasource)
 			if err != nil {
 				return nil, err
 			}
@@ -56,7 +56,7 @@ func (s *Service) ConfigureDatasource(ctx context.Context, organization Organiza
 			if strings.HasPrefix(currentDatasource.UID, datasourceUIDPrefix) {
 				// Delete the datasource as it is no longer desired
 				logger.Info("deleting datasource", "datasource", currentDatasource.UID)
-				err := s.deleteDatasource(ctx, currentDatasource.UID)
+				err := s.deleteDatasource(currentDatasource.UID)
 				if err != nil {
 					return nil, err
 				}
@@ -73,7 +73,7 @@ func (s *Service) ConfigureDatasource(ctx context.Context, organization Organiza
 		// We already took care of updating existing datasources ID in the previous loop
 		if desiredDatasource.ID == 0 {
 			logger.Info("creating datasource", "datasource", desiredDatasource.UID)
-			desiredDatasource, err = s.createDatasource(ctx, desiredDatasource)
+			desiredDatasource, err = s.createDatasource(desiredDatasource)
 			if err != nil {
 				return nil, err
 			}
@@ -89,26 +89,45 @@ func (s *Service) ConfigureDatasource(ctx context.Context, organization Organiza
 // generateDatasources generates the list of datasources for a given organization.
 // It configures the datasources to use the appropriate multi-tenant headers based on the organization's tenant IDs.
 // It returns the list of desired datasources.
-func (s *Service) generateDatasources(ctx context.Context, organization Organization) (datasources []Datasource) {
+func (s *Service) generateDatasources(organization Organization) (datasources []Datasource) {
 	// Multi-tenant header value is a pipe-separated list of tenant IDs
 	multiTenantIDsHeaderValue := strings.Join(organization.TenantIDs, "|")
 
 	// Add Loki datasource
-	datasources = append(datasources, DatasourceLoki().Merge(Datasource{
+	lokiDatasource := DatasourceLoki().Merge(Datasource{
 		Name: "Loki",
-		UID:  fmt.Sprintf("%sloki", datasourceUIDPrefix),
+		UID:  LokiDatasourceUID,
 		JSONData: map[string]any{
 			"httpHeaderName1": common.OrgIDHeader,
 		},
 		SecureJSONData: map[string]string{
 			"httpHeaderValue1": multiTenantIDsHeaderValue,
 		},
-	}))
+	})
+
+	// Add tracing integration if tracing is enabled
+	if s.cfg.Tracing.Enabled {
+		// Add derived fields for Loki to Tempo integration
+		lokiDatasource.JSONData["derivedFields"] = []map[string]any{
+			{
+				"name":          "traceID",
+				"matcherRegex":  "traceID=(\\w+)",
+				"datasourceUid": TempoDatasourceUID,
+				// Open a new tab when clicking the link
+				"targetBlank":     true,
+				"url":             "${__value.raw}",
+				"urlDisplayLabel": "Trace ID",
+			},
+		}
+
+	}
+
+	datasources = append(datasources, lokiDatasource)
 
 	// Add Mimir datasource
 	datasources = append(datasources, DatasourceMimir().Merge(Datasource{
 		Name:      "Mimir",
-		UID:       fmt.Sprintf("%smimir", datasourceUIDPrefix),
+		UID:       MimirDatasourceUID,
 		IsDefault: true,
 		JSONData: map[string]any{
 			"httpHeaderName1": common.OrgIDHeader,
@@ -121,7 +140,7 @@ func (s *Service) generateDatasources(ctx context.Context, organization Organiza
 	// Add Alertmanager datasource
 	datasources = append(datasources, DatasourceMimirAlertmanager().Merge(Datasource{
 		Name: "Mimir Alertmanager",
-		UID:  fmt.Sprintf("%smimir-alertmanager", datasourceUIDPrefix),
+		UID:  MimirAlertmanagerDatasourceUID,
 		JSONData: map[string]any{
 			"httpHeaderName1": common.OrgIDHeader,
 		},
@@ -129,6 +148,21 @@ func (s *Service) generateDatasources(ctx context.Context, organization Organiza
 			"httpHeaderValue1": multiTenantIDsHeaderValue,
 		},
 	}))
+
+	// Add Tempo datasource - only if tracing is enabled
+	if s.cfg.Tracing.Enabled {
+		datasources = append(datasources, DatasourceTempo().Merge(Datasource{
+			Name: "Tempo",
+			UID:  TempoDatasourceUID,
+			// TODO: Add multi-tenancy support
+			// JSONData: map[string]any{
+			// 	"httpHeaderName1": common.OrgIDHeader,
+			// },
+			// SecureJSONData: map[string]string{
+			// 	"httpHeaderValue1": multiTenantIDsHeaderValue,
+			// },
+		}))
+	}
 
 	if organization.Name == SharedOrg.Name {
 		// Add Mimir Cardinality datasources to the "Shared Org"
@@ -147,7 +181,7 @@ func (s *Service) generateDatasources(ctx context.Context, organization Organiza
 
 // createDatasource creates the given datasource in Grafana.
 // It returns the created datasource with its ID set.
-func (s *Service) createDatasource(ctx context.Context, datasource Datasource) (Datasource, error) {
+func (s *Service) createDatasource(datasource Datasource) (Datasource, error) {
 	created, err := s.grafanaClient.Datasources().AddDataSource(&models.AddDataSourceCommand{
 		UID:            datasource.UID,
 		Name:           datasource.Name,
@@ -174,7 +208,7 @@ func (s *Service) createDatasource(ctx context.Context, datasource Datasource) (
 // updateDatasource updates the given datasource in Grafana.
 // The datasource is identified by its UID.
 // It returns the updated datasource with its ID set.
-func (s *Service) updateDatasource(ctx context.Context, datasource Datasource) (Datasource, error) {
+func (s *Service) updateDatasource(datasource Datasource) (Datasource, error) {
 	resp, err := s.grafanaClient.Datasources().UpdateDataSourceByUID(datasource.UID, &models.UpdateDataSourceCommand{
 		UID:            datasource.UID,
 		Name:           datasource.Name,
@@ -200,7 +234,7 @@ func (s *Service) updateDatasource(ctx context.Context, datasource Datasource) (
 
 // deleteDatasource deletes the datasource with the given UID.
 // If the datasource does not exist, no error is returned.
-func (s *Service) deleteDatasource(ctx context.Context, uid string) error {
+func (s *Service) deleteDatasource(uid string) error {
 	_, err := s.grafanaClient.Datasources().DeleteDataSourceByUID(uid)
 	if err != nil {
 		var notFound *datasources.DeleteDataSourceByUIDNotFound